+++ Test/Integration/Model/FormRepositoryTest.php
@@ -1,17 +1,23 @@
 <?php
-
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
 namespace Alekseon\CustomFormsBuilder\Test\Integration\Model;
 
 use PHPUnit\Framework\TestCase;
 use Magento\TestFramework\Helper\Bootstrap;
 use Alekseon\CustomFormsBuilder\Model\FormRepository;
 
+/**
+ * Class FormRepositoryTest
+ * @package Alekseon\CustomFormsBuilder\Test\Integration\Model
+ */
 class FormRepositoryTest extends TestCase
 {
-
     private $formRepository;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $objectManager = Bootstrap::getObjectManager();